restructured chat field
