+++ 1 - Students/ArtemKarlov/src/components/ChatField/ChatField.jsx
@@ -3,39 +3,23 @@ import React, { Fragment } from 'react';
 
 import {connect} from 'react-redux';
 
-import { makeStyles } from '@material-ui/core/styles';
-import Avatar from '@material-ui/core/Avatar';
+import ChatPlaceHolder from '../ChatPlaceHolder/ChatPlaceHolder.jsx';
+import ChatDailog from '../ChatDialog/ChatDialog.jsx';
 
-const useStyles = makeStyles(() => ({
-    accountAvatar: {width: '100%', height: '100%'}
-}));
-
-// import ChatInfo from '../ChatInfo/ChatInfo.jsx';
-import Messages from '../Messages/Messages.jsx';
-
-function ChatField(props) {
-    const classes = useStyles();
-    const {chatId, chats} = props;
-    const currentChat = chats.find((chat) => chat.id === chatId);
+export default function ChatField(props) {
+    const {chatId} = props;    
+    
     return (
-        <Fragment>
-            <section className="lyaout__chat chat">
-                <div className="chat__header chat-header">
-                    <p className="chat-header__label">Chat with</p>
-                    <div className="chat-header__details">
-                        <div className="chat-header__img img-container ">
-                            <Avatar alt={currentChat.title} src={currentChat.avatarUrl} className={classes.accountAvatar}/>
-                        </div>                    
-                        <h2 className="chat-header__title">{currentChat.title}</h2>
-                    </div>                    
-                </div>
-                <Messages chatId={chatId} />
-            </section>
-        </Fragment>
+        <section className="lyaout__chat chat">
+            {(chatId === 'init') ?
+                <ChatPlaceHolder /> : 
+                <ChatDailog chatId={chatId} />
+            }         
+        </section>
     );
 }
 
-const mapStateToProps = ({chatsReducer}) => ({
-    chats: chatsReducer.chats,
-});
-export default connect(mapStateToProps)(ChatField);
+// const mapStateToProps = ({chatsReducer}) => ({
+//     chats: chatsReducer.chats,
+// });
+// export default connect(mapStateToProps)(ChatField);
