+++ 1 - Students/ArtemKarlov/src/components/ChatField/style.css
@@ -9,53 +9,3 @@
     flex-direction: column;
     justify-content: space-between;
 }
-
-.chat__header {
-    flex-shrink: 0;
-}
-
-.chat-header {
-    padding-top: 20px;
-    padding-bottom: 20px;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
-}
-
-.chat-header::after {
-    content: '';
-    display: block;
-    width: 70%;
-    position: absolute;
-    bottom: 0px;
-    align-self: center;
-
-    opacity: 0.1;
-    border-bottom: 1px solid #FFFFFF; 
-}
-
-.chat-header__label {
-    margin-bottom: 12px;
-    font-size: 12px;
-    opacity: 0.3;
-}
-
-.chat-header__details {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-}
-
-.chat-header__img {
-    width: 55px;
-    height: 55px;
-    margin-right: 20px;
-}
-
-.chat-header__title {
-    font-weight: bold;
-    font-size: 24px;
-}
