+++ 1 - Students/ArtemKarlov/src/components/ChatPlaceHolder/ChatPlaceHolder.jsx
@@ -1,0 +1,14 @@
+import './style.css';
+import React, { Fragment } from 'react';
+
+
+export default function ChatPlaceHolder(props) {
+
+    return (
+        <div className="chat__placeholder">
+            <p>Please select a chat to start messaging</p>
+        </div>
+        
+    );
+}
+
