+++ 1 - Students/ArtemKarlov/src/components/ChatPlaceHolder/style.css
@@ -1,0 +1,10 @@
+.chat__placeholder {
+    padding: 10px 20px;
+    width: 70%;
+    margin: auto;
+    background-color: rgba(37, 140, 96, 0.4);
+    border-radius: 8px;
+    text-align: center;
+    
+}
+
